extract getTagFilters
