+++ packages/components/src/templates/next/layouts/Collection/filterUtils/index.ts
@@ -1,3 +1,4 @@
 export * from "./getCategoryFilters"
 export * from "./getYearFilters"
+export * from "./getTagFilters"
 export * from "./constants"
